+++ src/Client/Exception/HandlerNotFound.php
@@ -1,21 +1,22 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\API\Client\Exception;
 
 use Psr\Http\Message\RequestInterface;
 
 /**
- * Class HandlerNotFound
- *
- * @package DigipolisGent\API\Client\Exception
+ * Exception thrown when no handler is found for the given request.
  */
 class HandlerNotFound extends Handler
 {
     /**
      * @param RequestInterface $request
-     * @return HandlerNotFound
+     *
+     * @return \DigipolisGent\API\Client\Exception\HandlerNotFound
      */
-    public static function fromRequest(RequestInterface $request)
+    public static function fromRequest(RequestInterface $request): HandlerNotFound
     {
         return new self(sprintf('No handler was registered for %s', get_class($request)));
     }