+++ src/Client/Exception/Request.php
@@ -1,13 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\API\Client\Exception;
 
 /**
- * Client Request exception.
- *
- * @package DigipolisGent\API\Client\Exception
+ * Exception related to a request.
  */
-class Request extends \Exception
+class Request extends Exception
 {
 
 }